Print the memory map given by the bootloader and fix issues with serial_send_value

+++ include/kernel/serial.h
@@ -2,7 +2,7 @@
 #define _SERIAL_H_
 
 void serial_init();
-void serial_send_value  (int port, int base, int value);
+void serial_send_value  (int port, u8 base, u32 value);
 void serial_send_string (int port, char* str);
 void serial_send_char   (int port, char c);
