+++ include/typedef.h
@@ -2,10 +2,10 @@
 #define _TYPEDEF_H_
 
 typedef enum bool bool;
-typedef unsigned char  u8;
-typedef unsigned short u16;
-typedef unsigned int   u32;
-typedef unsigned long  u64;
+typedef unsigned char  		u8;
+typedef unsigned short 		u16;
+typedef unsigned int   		u32;
+typedef unsigned long long	u64;
 
 enum bool
 {
